+++ .travis.yml
@@ -10,4 +10,4 @@ before_install:
 script:
 - go vet ./...
 - $HOME/gopath/bin/golint ./...
-- $HOME/gopath/bin/goveralls -service=travis-ci
+- $HOME/gopath/bin/goveralls -service=travis-ci ./...
