+++ ovh/configuration_test.go
@@ -1,4 +1,4 @@
-package govh
+package ovh
 
 import (
 	"io/ioutil"
